fix #165  - Fixed a bug that MusicFolder path is not analyzed correctly.  - Added test for each Analyzer/Query /Service processing hierarchy because there were no legacy test cases
