+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/service/search/QueryFactory.java
@@ -24,7 +24,6 @@ import org.airsonic.player.domain.MediaFile.MediaType;
 import org.airsonic.player.domain.MusicFolder;
 import org.airsonic.player.domain.RandomSearchCriteria;
 import org.airsonic.player.domain.SearchCriteria;
-import org.apache.lucene.analysis.Analyzer;
 import org.apache.lucene.document.IntPoint;
 import org.apache.lucene.index.Term;
 import org.apache.lucene.queryparser.classic.MultiFieldQueryParser;
@@ -34,18 +33,11 @@ import org.apache.lucene.search.BooleanClause.Occur;
 import org.apache.lucene.search.BooleanQuery;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.TermQuery;
-import org.apache.lucene.search.spans.SpanOrQuery;
-import org.apache.lucene.search.spans.SpanQuery;
-import org.apache.lucene.search.spans.SpanTermQuery;
-import org.apache.lucene.util.BytesRef;
-import org.apache.lucene.util.NumericUtils;
 import org.slf4j.LoggerFactory;
 
 import java.util.Arrays;
 import java.util.List;
 
-import static com.tesshu.jpsonic.service.search.IndexType.ALBUM_ID3;
-import static com.tesshu.jpsonic.service.search.IndexType.ARTIST_ID3;
 import static com.tesshu.jpsonic.service.search.IndexType.normalizeGenre;
 import static org.springframework.util.ObjectUtils.isEmpty;
 
@@ -71,10 +63,9 @@ public class QueryFactory {
 		String[] targetFields = Arrays.stream(indexType.getFields())
 				.filter(f -> !f.equals(FieldNames.FOLDER))
 				.toArray(s -> new String[s]);
-
-		Analyzer analyzer = AnalyzerFactory.getInstance().getQueryAnalyzer(criteria.getQuery());
-
-		MultiFieldQueryParser parser = new MultiFieldQueryParser(targetFields, analyzer);
+		
+		MultiFieldQueryParser parser =
+		    new MultiFieldQueryParser(targetFields, AnalyzerFactory.getInstance().getAnalyzer());
 
 		BooleanQuery.Builder builder = new BooleanQuery.Builder();
 		try {
@@ -83,21 +74,16 @@ public class QueryFactory {
 			LoggerFactory.getLogger(QueryFactory.class).error("Error during query analysis.", e);
 		}
 
-		SpanQuery[] musicFolderQueries = musicFolders.stream().map(musicFolder -> {
-			if (indexType == ALBUM_ID3 || indexType == ARTIST_ID3) {
-				byte[] bytes = new byte[Integer.BYTES];
-				NumericUtils.intToSortableBytes(musicFolder.getId(), bytes, 0);
-				BytesRef ref = new BytesRef(bytes);
-				return new SpanTermQuery(new Term(FieldNames.FOLDER_ID, ref));
-			} else {
-				return new SpanTermQuery(
-						new Term(FieldNames.FOLDER, musicFolder.getPath().getPath()));
-			}
-		}).toArray(i -> new SpanQuery[i]);
+    BooleanQuery.Builder sub = new BooleanQuery.Builder();
+    musicFolders.forEach(musicFolder -> {
+      if (indexType == IndexType.ALBUM_ID3 || indexType == IndexType.ARTIST_ID3) {
+        sub.add(new TermQuery(new Term(FieldNames.FOLDER_ID, musicFolder.getId().toString())), Occur.SHOULD);
+      } else {
+        sub.add(new TermQuery(new Term(FieldNames.FOLDER, musicFolder.getPath().getPath())), Occur.SHOULD);
+      }
+    });
 
-		builder.add(new SpanOrQuery(musicFolderQueries), Occur.MUST);
-
-		return builder.build();
+    return builder.add(sub.build(), Occur.MUST).build();
 
 	}
 
@@ -111,9 +97,6 @@ public class QueryFactory {
 					new Term(FieldNames.GENRE, normalizeGenre.apply(criteria.getGenre()))), Occur.MUST);
 		}
 
-		/*
-		 * @see IntPoint#newRangeQuery
-		 */
 		if(!(isEmpty(criteria.getFromYear()) && isEmpty(criteria.getToYear()))) {
 			builder.add(IntPoint.newRangeQuery(
 					FieldNames.YEAR,
@@ -122,42 +105,44 @@ public class QueryFactory {
 					Occur.MUST);
 		}
 
-		SpanQuery[] musicFolderQueries = criteria.getMusicFolders().stream()
-				.map(musicFolder -> new SpanTermQuery(new Term(FieldNames.FOLDER, musicFolder.getPath().getPath())))
-				.toArray(i -> new SpanQuery[i]);
-		builder.add(new SpanOrQuery(musicFolderQueries), Occur.MUST);
+		BooleanQuery.Builder sub = new BooleanQuery.Builder();
+		criteria.getMusicFolders().forEach(musicFolder -> 
+		sub.add(new TermQuery(new Term(
+		      FieldNames.FOLDER,
+		      musicFolder.getPath().getPath())),
+		      Occur.SHOULD));
 
-		return builder.build();
+		return builder.add(sub.build(), Occur.MUST).build();
 	}
 
 	public static Query searchByName(String name, String field) {
-		Analyzer queryAnalyzer = AnalyzerFactory.getInstance().getQueryAnalyzer(name);
-		QueryParser queryParser = new QueryParser(field, queryAnalyzer);
+		QueryParser queryParser = new QueryParser(field, AnalyzerFactory.getInstance().getAnalyzer());
+		queryParser.setDefaultOperator(QueryParser.Operator.AND);
 		Query parsedQuery = null;
-		try {
-			parsedQuery = queryParser.parse(name);
-		} catch (ParseException e) {
-			LoggerFactory.getLogger(QueryFactory.class).error("Error during query analysis.", e);
-		}
+		    parsedQuery = queryParser.createBooleanQuery(field, QueryParser.escape(name), Occur.MUST);
 		return parsedQuery;
 	}
 
-	public static Query searchByNameMusicFolderPath(List<MusicFolder> musicFolders) {
-		SpanQuery[] musicFolderQueries = musicFolders.stream()
-				.map(musicFolder -> new SpanTermQuery(new Term(FieldNames.FOLDER, musicFolder.getPath().getPath())))
-				.toArray(i -> new SpanQuery[i]);
-		return new SpanOrQuery(musicFolderQueries);
+	public static Query searchRandomAlbum(List<MusicFolder> musicFolders) {
+    BooleanQuery.Builder builder = new BooleanQuery.Builder();
+    BooleanQuery.Builder sub = new BooleanQuery.Builder();
+    musicFolders.forEach(musicFolder -> 
+    sub.add(new TermQuery(new Term(
+          FieldNames.FOLDER,
+          musicFolder.getPath().getPath())),
+          Occur.SHOULD));
+    return builder.add(sub.build(), Occur.MUST).build();
 	}
-	
-	public static Query searchByNameMusicFolderIds(List<MusicFolder> musicFolders) {
-		SpanQuery[] musicFolderQueries = musicFolders.stream()
-				.map(musicFolder -> {
-					byte[] bytes = new byte[Integer.BYTES];
-					NumericUtils.intToSortableBytes(musicFolder.getId(), bytes, 0);
-					BytesRef ref = new BytesRef(bytes);
-					return new SpanTermQuery(new Term(FieldNames.FOLDER_ID, ref));
-				}).toArray(i -> new SpanQuery[i]);
-		return new SpanOrQuery(musicFolderQueries);
+
+	public static Query searchRandomAlbumId3(List<MusicFolder> musicFolders) {
+	  BooleanQuery.Builder builder = new BooleanQuery.Builder();
+	  BooleanQuery.Builder sub = new BooleanQuery.Builder();
+	  musicFolders.forEach(musicFolder -> 
+	    sub.add(new TermQuery(new Term(
+	        FieldNames.FOLDER_ID,
+	        musicFolder.getId().toString())),
+	        Occur.SHOULD));
+	    return builder.add(sub.build(), Occur.MUST).build();
 	}
 
 }
